Replace id_setter with root_validator so that id is assigned properly
Fixes #87

The reason why id_setter was not working was because pydantic validates fields in order. And when using the validator decorator, all validated fields go in the values argument. (See pydantic docs on validators: https://pydantic-docs.helpmanual.io/usage/validators/)

However, this causes an issue for us since id is the first field in any model, and so cannot rely on any other field when validated. To fix this issue, we can just use a plain old pydantic root_validator, which gives us access to the entire model's data.