Add ability to initialise a MongoDB Connection from a config file.
Add a new module, which contains an interface for setting up and creating a MongoDB connection from a .INI structured config file.

The structure of the .INI expects a MongoDB connection string formatted according to the official documentation: 
https://docs.mongodb.com/manual/reference/connection-string/

The implementation takes care of possible encoding issues the end user may have when creating a connection string, 
such as using percent encoding for usernames and passwords.