Added functionality to download database entries into files.
The system is designed in a sort of reverse way to how the readers work.

The main difference is how the standard writers method work, as instead of reading from a file, we had to implementent a way to create a file and place information into the created file.

Other than that, the other files filewriter.py and writers.py are basically the same as there reading counterparts.

Changes were also made to init.py to include the get_header function