Use the database config system to automatically connect to a database
Remove the database argument for functions that upload/download the database. As the database parameter is now deduced using the config system.

Using the new config system (#89), we can automatically set up and initialise a database connection. 
Since this connection is most likely to be the same, it doesn't make much sense to pass in the database parameter every time we want to use the database.

Therefore, this method is more concise and easy to use.