Change BaseModel to validate assignments
By default, pydantic only validates data when the document is first constructed.
The issue with this is that the user could potentially directly assign fields to invalid values, since pydantic does not validate assignments by default.
Fortunately, pydantic already has this functionality built in, we just need to enable it by adding a field to the Config class.